requirements.txt
asgiref==3.2.10
Django==3.1
django-email-verification==0.0.7
django-filter==2.3.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1